src/requirements/base.txt
# Core
Django>=1.8,<1.9
psycopg2>=2.6,<2.7

# Packages
ipython
django-extensions
djangorestframework>=3.2,<3.3
drf-nested-routers
Werkzeug
